user service , repository and controller set and connection done

+++ instapostBusinesslayer/Interface/UserInterface/UserIRepository.cs
@@ -4,7 +4,7 @@ namespace instapostBusinesslayer.Interface.UserInterface
 {
     public interface UserIRepository
     {
-        Task<int> CreateUser(UserEntity ue);
+        Task<long> CreateUser(UserEntity ue);
         Task<UserEntity> GetUserById(long id);
 
         Task<bool> UpdateUser(UserEntity ue);
